+++ CHANGELOG.md
@@ -36,6 +36,15 @@
 ### Under the Hood
 - Upgrade GeographicLib to version 2.2 (from version 2.1.2)
 
+## 0.16.3
+
+### Improvements
+- Add an exception handler dialog that provides information about abnormal program termination, including
+  * Information about the exception
+  * Stack trace (using the external Cpptrace library v0.1.1)
+  * Application- and basic system information
+  * Unix signal handling (macOS, Linux only)
+
 ## 0.16.2
 
 ### Bug Fixes